pyproject.toml
[project]
name = "mcq"
version = "1.0.0"
description = "A Rich-based terminal UI for answering multiple choice questions from YAML files"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.scripts]
mcq = "mcq_tui:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
py-modules = ["mcq_tui"]

